enterprise_doc_qa_full_v3/requirements.txt
streamlit
sentence-transformers
faiss-cpu
python-docx
pymupdf
openai